+++ app/Http/Resources/AssetResource.php
@@ -18,9 +18,12 @@ class AssetResource extends JsonResource
             'created_at' => $this->created_at,
             'updated_at' => $this->updated_at,
             'status' => $this->status,
+            'custom_id' => $this->custom_id,
             'category' => new CategoryResource($this->whenLoaded('category')),
             'location' => new LocationResource($this->whenLoaded('location')),
             'tags' => TagResource::collection($this->whenLoaded('tags')),
+            'custom_field_values' => $this->whenLoaded('customFieldValues'),
+            'additional_files' => $this->whenLoaded('additionalFiles')
         ];
     }
 }