+++ app/Models/Asset.php
@@ -6,6 +6,7 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Concerns\HasUlids;
 
 class Asset extends Model
@@ -20,7 +21,8 @@ class Asset extends Model
         'team_id',
         'category_id',
         'location_id',
-        'status'
+        'status',
+        'custom_id'
     ];
 
     public function team(): BelongsTo
@@ -46,5 +48,17 @@ class Asset extends Model
     public function kit(): BelongsTo
     {
         return $this->belongsTo(Kit::class);
+    }
+
+    public function customFieldValues(): HasMany
+    {
+        return $this->hasMany(AssetCustomFieldValue::class);
+    }
+
+    public function additionalFiles(): BelongsToMany
+    {
+        return $this->belongsToMany(AdditionalFile::class, 'asset_additional_file', 'asset_id', 'additional_file_id')
+                    ->using(AssetAdditionalFile::class)
+                    ->withTimestamps();
     }
 }